(Chore): Rename GitHub instances to Git (#7625)
