+++ frontend/src/hooks/query/use-search-repositories.ts
@@ -4,7 +4,7 @@ import OpenHands from "#/api/open-hands";
 export function useSearchRepositories(query: string) {
   return useQuery({
     queryKey: ["repositories", query],
-    queryFn: () => OpenHands.searchGitHubRepositories(query, 3),
+    queryFn: () => OpenHands.searchGitRepositories(query, 3),
     enabled: !!query,
     select: (data) => data.map((repo) => ({ ...repo, is_public: true })),
     staleTime: 1000 * 60 * 5, // 5 minutes
